# Conflicts:
#	Sneaker/Controllers/CartController.cs
#	Sneaker/Repository/CartRepo.cs
#	Sneaker/Repository/Interface/ICartRepo.cs
#	Sneaker/ViewModel/CartViewModel.cs
#	Sneaker/Views/Cart/Index.cshtml
#	Sneaker/Views/Cart/SubmitOrder.cshtml

+++ Sneaker/Repository/CartRepo.cs
@@ -1,6 +1,5 @@
 ﻿using Microsoft.AspNetCore.Http;
 using Microsoft.EntityFrameworkCore;
-using Microsoft.Extensions.Configuration;
 using Sneaker.Data;
 using Sneaker.Helpers;
 using Sneaker.Models;
@@ -9,6 +8,7 @@ using Sneaker.ViewModel;
 using System;
 using System.Collections.Generic;
 using System.Linq;
+using System.Security.Claims;
 using System.Threading.Tasks;
 
 namespace Sneaker.Repository
@@ -17,12 +17,10 @@ namespace Sneaker.Repository
     {
         private readonly ApplicationDbContext _dbContext;
         private IHttpContextAccessor _httpContextAccessor;
-        private readonly IConfiguration _configuration;
-        public CartRepo(ApplicationDbContext dbContext, IHttpContextAccessor httpContextAccessor, IConfiguration configuration)
+        public CartRepo(ApplicationDbContext dbContext, IHttpContextAccessor httpContextAccessor)
         {
             _dbContext = dbContext;
             _httpContextAccessor = httpContextAccessor;
-            _configuration = configuration;
         }
 
         public Cart cart(string userId)
@@ -66,15 +64,24 @@ namespace Sneaker.Repository
         }
         public int GetCount(string userId)
         {
-            return _dbContext.Carts.Where(c => c.UserId == userId).Sum(c => c.Quantity);
+            return _dbContext.Carts.Where(c => c.UserId == userId).Count();
         }
 
         public bool RemoveCart(int id)
         {
             var CartItem = _dbContext.Carts.SingleOrDefault(c => c.Products.Id == id);
+            var localAmount = 0;
             if (CartItem != null)
             {
-                _dbContext.Carts.Remove(CartItem);
+                if (CartItem.Quantity > 1)
+                {
+                    CartItem.Quantity--;
+                    localAmount = CartItem.Quantity;
+                }
+                else
+                {
+                    _dbContext.Carts.Remove(CartItem);
+                }
             }
             _dbContext.SaveChanges();
             return true;
@@ -82,43 +89,67 @@ namespace Sneaker.Repository
 
         public bool ClearCart(int id)
         {
-            var cartItems = _dbContext.Carts.Find(id);
+            var cartItems = _dbContext.Carts.Where(c => c.Id == id);
             _dbContext.Carts.RemoveRange(cartItems);
             _dbContext.SaveChanges();
             return true;
         }
 
-        public bool CreateOrder(CartViewModel invoiceVM, string userId)
+        public bool CreateOrder(Invoice invoice, string userId)
         {
-            invoiceVM.Invoices.CreateAt = DateTime.Now;
-            invoiceVM.Invoices.OwnerId = userId;
-            _dbContext.Invoice.Add(invoiceVM.Invoices);
-            _dbContext.SaveChanges();
+            invoice.CreateAt = DateTime.Now;
+            invoice.OwnerId = userId;
+            _dbContext.Invoice.Add(invoice);
             decimal orderTotal = 0;
             var cartItems = GetCartItem(userId);
             foreach (var item in cartItems)
             {
-                var invoiceDetail = new InvoiceDetails
+                var orderDetail = new InvoiceDetails
                 {
                     ProductId = item.Products.Id,
-                    InvoiceId = invoiceVM.Invoices.Id,
-                    Price = item.Products.Price * item.Quantity,
+                    InvoiceId = invoice.Id,
+                    Price = item.Products.Price,
                     Quantity = item.Quantity,
                     UserId = item.UserId,
                 };
                 orderTotal += (item.Quantity * item.Products.Price);
-                _dbContext.InvoiceDetails.Add(invoiceDetail);
+                _dbContext.InvoiceDetails.Add(orderDetail);
             }
-            invoiceVM.Invoices.OrderTotal = orderTotal;
+            invoice.OrderTotal = orderTotal;
             _dbContext.SaveChanges();
             return true;
         }
-
-        public async Task<bool> SubmitOrder(string paymentId, string payerId, CartViewModel invoiceVM)
-        {
-            var paypalAPI = new PayPalAPI(_configuration);
-            await paypalAPI.ExecutedPayment(paymentId, payerId); 
-            return true;
-        }
+        //public bool CreateOrder(Invoice invoice, string userId)
+        //{
+        //    var cartItems = GetCartItem(userId);
+        //    foreach (var item in cartItems)
+        //    {
+        //        var newOrder = new Invoice
+        //        {
+        //            FirstName = invoice.FirstName,
+        //            LastName = invoice.LastName,
+        //            Email = invoice.Email,
+        //            PhoneNumber = invoice.PhoneNumber,
+        //            Address = invoice.Address,
+        //            State = invoice.State,
+        //            Country = invoice.Country,
+        //            PostalCode = invoice.PostalCode,
+        //            OrderTotal = (item.Quantity * item.Products.Price),
+        //            OwnerId = userId
+        //        };
+        //        _dbContext.Invoice.Add(newOrder);
+        //        var newDetails = new InvoiceDetails
+        //        {
+        //            ProductId = item.Products.Id,
+        //            InvoiceId = invoice.Id,
+        //            Price = item.Products.Price,
+        //            Quantity = item.Quantity,
+        //            UserId = item.UserId,
+        //        };
+        //        _dbContext.InvoiceDetails.Add(newDetails);
+        //    }
+        //    _dbContext.SaveChanges();
+        //    return true;
+        //}
     }
 }
